fix: Maximum Length Sub-Itemのlengthの型をu16に修正

+++ src/dicom-server/src/dicom/network/pdu/a_associate/items/user_information/sub_items/maximum_length.rs
@@ -13,7 +13,7 @@ impl MaximumLength {
         8
     }
 
-    pub fn length(&self) -> u32 {
+    pub fn length(&self) -> u16 {
         4
     }
